rename enumarble file match snake case
